+++ gz/ai_auto_learn_config.php
@@ -5,16 +5,18 @@
 return array (
   'enabled' => true,
   'interval_hours' => 4,
+  'target_mode' => 'all',
   'target_sites' => 
   array (
     0 => '如意',
+    1 => '量子',
   ),
   'play_from_patterns' => 
   array (
     0 => 'rym3u8',
   ),
-  'videos_per_site' => 1,
-  'max_sites_per_run' => 1,
+  'videos_per_site' => 50,
+  'max_sites_per_run' => 3,
   'min_segments' => 50,
   'max_ad_percentage' => 90,
   'max_exec_time_per_video' => 30,
@@ -23,4 +25,10 @@ return array (
   'min_learn_count_to_track' => 1,
   'access_key' => '',
   'last_run_time' => '2026-08-03 03:21:52',
+  'auto_trigger_on_request' => true,
+  'auto_cleanup_stale_rules' => true,
+  'stale_rule_days' => 30,
+  'cleanup_health_timeout' => 6,
+  'cleanup_interval_hours' => 24,
+  'last_cleanup_time' => '2026-08-03 03:32:30',
 );